Don't expose getSessionExpiry
