+++ public/app/core/services/backend_srv.ts
@@ -20,6 +20,7 @@ import { GrafanaEdition } from '@grafana/data/src/types/config';
 import { BackendSrv as BackendService, BackendSrvRequest, config, FetchError, FetchResponse } from '@grafana/runtime';
 import appEvents from 'app/core/app_events';
 import { getConfig } from 'app/core/config';
+import { getSessionExpiry } from 'app/core/utils/auth';
 import { loadUrlToken } from 'app/core/utils/urlToken';
 import { DashboardModel } from 'app/features/dashboard/state';
 import { DashboardSearchItem } from 'app/features/search/types';
@@ -391,7 +392,7 @@ export class BackendSrv implements BackendService {
 
                 let authChecker = this.loginPing();
 
-                const expired = contextSrv.getSessionExpiry() * 1000 < Date.now();
+                const expired = getSessionExpiry() * 1000 < Date.now();
                 if (config.featureToggles.clientTokenRotation && expired) {
                   authChecker = this.rotateToken();
                 }
